netstack告警清理

+++ utils/napi_utils/src/event_listener.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2022 Huawei Device Co., Ltd.
+ * Copyright (c) 2021-2024 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -164,7 +164,7 @@ uint64_t GetCurrentThreadId()
 #if HAS_NETMANAGER_BASE
     auto tid = ffrt_this_task_get_id();
     if (tid == 0) {
-        tid = syscall(SYS_gettid);
+        tid = static_cast<uint64_t>(syscall(SYS_gettid));
     }
     return tid;
 #else
